Don't encode the authenticated model in state
This simplifies auth and allows it to focus on what it does well,
but has the downside of allowing an authenticated model to exist
in an unauthenticated state. This means you have to explicitly clear
it when you sign out, as shown in the example.

I'm ok with this, however because I expect it to work well with modem
where a modem type holds the authenticated data and you match on both
auth state and modem in the router.

Will add an example of this shortly to test my theory.